# Improve performance of FormattingHelper.CountDigits(uint)