--- a/src/libraries/System.Private.CoreLib/src/System/Number.Formatting.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Number.Formatting.cs
@@ -1593,14 +1593,13 @@ namespace System
 
         internal static unsafe string UInt32ToDecStr(uint value)
         {
-            // Intrinsified in mono interpreter
-            int bufferLength = FormattingHelpers.CountDigits(value);
-
             // For single-digit values that are very common, especially 0 and 1, just return cached strings.
-            if (bufferLength == 1)
+            if (value < 10)
             {
                 return s_singleDigitStringCache[value];
             }
+
+            int bufferLength = FormattingHelpers.CountDigits(value);
 
             string result = string.FastAllocateString(bufferLength);
             fixed (char* buffer = result)
@@ -1935,14 +1934,13 @@ namespace System
 
         internal static unsafe string UInt64ToDecStr(ulong value)
         {
-            // Intrinsified in mono interpreter
-            int bufferLength = FormattingHelpers.CountDigits(value);
-
             // For single-digit values that are very common, especially 0 and 1, just return cached strings.
-            if (bufferLength == 1)
+            if (value < 10)
             {
                 return s_singleDigitStringCache[value];
             }
+
+            int bufferLength = FormattingHelpers.CountDigits(value);
 
             string result = string.FastAllocateString(bufferLength);
             fixed (char* buffer = result)