@stephentoub 2022-10-03
I'll merge this and follow-up on the 64-bit case.